Adopt solidus_dev_support and migrate the name to Solidus
