+++ .rspec
@@ -1,1 +1,2 @@
 --color
+--require spec_helper
