+++ .rubocop.yml
@@ -1,12 +1,2 @@
-inherit_from: .rubocop_todo.yml
-
 require:
-  - solidus_extension_dev_tools/rubocop
-
-inherit_gem:
-  solidus_extension_dev_tools: .rubocop.extension.yml
-
-AllCops:
-  Exclude:
-    - spec/dummy/**/*
-    - vendor/**/*
+  - solidus_dev_support/rubocop
