+++ README.md
@@ -1,5 +1,5 @@
-Spree::HandlingFees
-=================
+SolidusHandlingFees
+===================
 
 [![CircleCI](https://circleci.com/gh/solidusio-contrib/solidus_handling_fees.svg?style=svg)](https://circleci.com/gh/solidusio-contrib/solidus_handling_fees)
 
@@ -8,21 +8,21 @@ Adds Handling Fee functionality to [Spree](http://github.com/spree/spree).
 Installation
 ------------
 
-Add spree_handling_fees to your Gemfile:
+Add solidus_handling_fees to your Gemfile:
 
 ```ruby
-gem 'spree_handling_fees', github: 'spree-contrib/spree_handling_fees'
+gem 'solidus_handling_fees', github: 'solidusio-contrib/solidus_handling_fees'
 ```
 
 Bundle your dependencies and run the installation generator:
 
 ```shell
 bundle
-bundle exec rails g spree_handling_fees:install
+bundle exec rails g solidus_handling_fees:install
 ```
 
 Usage
--------
+-----
 
 To create a handling fee, edit a Stock Location in the admin and apply a calculator. For example, you might use the Flat Rate calculator to set a flat fee of $3.99 on all shipments.
 
@@ -31,7 +31,7 @@ If you'd like to specify only certain orders to have a handling fee, override th
 Testing
 -------
 
-First bundle your dependencies, then run `rake`. `rake` will default to building the dummy app if it does not exist, then it will run specs. The dummy app can be regenerated by using `rake test_app`.
+First bundle your dependencies, then run `rake`. `rake` will default to building the dummy app if it does not exist, then it will run specs, and [Rubocop](https://github.com/bbatsov/rubocop) static code analysis. The dummy app can be regenerated by using `rake test_app`.
 
 ```shell
 bundle
@@ -42,7 +42,16 @@ When testing your application's integration with this extension you may use its 
 Simply add this require statement to your spec_helper:
 
 ```ruby
-require 'spree_handling_fees/factories'
+require 'solidus_handling_fees/factories'
+```
+
+Releasing
+---------
+
+Your new extension version can be released using `gem-release` like this:
+
+```shell
+bundle exec gem bump -v VERSION --tag --push --remote upstream && gem release
 ```
 
 Copyright (c) 2014 Brian Christian, released under the GPL V3.
