+++ Rakefile
@@ -1,23 +1,6 @@
 # frozen_string_literal: true
 
-require 'bundler'
-Bundler::GemHelper.install_tasks
+require 'solidus_dev_support/rake_tasks'
+SolidusDevSupport::RakeTasks.install
 
-require 'rspec/core/rake_task'
-require 'spree/testing_support/extension_rake'
-
-RSpec::Core::RakeTask.new
-
-task :default do
-  if Dir["spec/dummy"].empty?
-    Rake::Task[:test_app].invoke
-    Dir.chdir("../../")
-  end
-  Rake::Task[:spec].invoke
-end
-
-desc 'Generates a dummy app for testing'
-task :test_app do
-  ENV['LIB_NAME'] = 'spree_handling_fees'
-  Rake::Task['extension:test_app'].invoke
-end
+task default: 'extension:specs'
