+++ lib/generators/solidus_handling_fees/install/install_generator.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+module SolidusHandlingFees
+  module Generators
+    class InstallGenerator < Rails::Generators::Base
+      class_option :auto_run_migrations, type: :boolean, default: false
+
+      def add_migrations
+        run 'bin/rails railties:install:migrations FROM=solidus_handling_fees'
+      end
+
+      def run_migrations
+        run_migrations = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask('Would you like to run the migrations now? [Y/n]'))
+        if run_migrations
+          run 'bin/rails db:migrate'
+        else
+          puts 'Skipping bin/rails db:migrate, don\'t forget to run it!' # rubocop:disable Rails/Output
+        end
+      end
+    end
+  end
+end
