+++ lib/solidus_handling_fees.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'solidus_core'
+require 'solidus_support'
+
+require 'solidus_handling_fees/version'
+require 'solidus_handling_fees/engine'
+
+# Supprt the legacy constant name.
+SpreeHandlingFees = SolidusHandlingFees
