+++ lib/solidus_handling_fees/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module SolidusHandlingFees
+  VERSION = '2.4.0'
+end
