+++ lib/spree_handling_fees.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
-require 'spree_core'
-require 'spree_handling_fees/engine'
-require 'deface'
+warn \
+  'DEPRECATED REQUIRE: Please update your refs from "spree_handling_fees" ' \
+  'to "solidus_handling_fees".', uplevel: 1
+require 'solidus_handling_fees'
