+++ lib/spree_handling_fees/engine.rb
@@ -1,22 +1,6 @@
 # frozen_string_literal: true
 
-module SpreeHandlingFees
-  class Engine < Rails::Engine
-    require 'spree/core'
-    isolate_namespace Spree
-    engine_name 'spree_handling_fees'
-
-    # use rspec for tests
-    config.generators do |g|
-      g.test_framework :rspec
-    end
-
-    def self.activate
-      Dir.glob(File.join(File.dirname(__FILE__), '../../app/**/*_decorator*.rb')) do |c|
-        Rails.configuration.cache_classes ? require(c) : load(c)
-      end
-    end
-
-    config.to_prepare &method(:activate).to_proc
-  end
-end
+warn \
+  'DEPRECATED REQUIRE: Please update your refs from "spree_handling_fees" ' \
+  'to "solidus_handling_fees".', uplevel: 1
+require 'solidus_handling_fees'
